vcs/canvexoverlay: Do not use pre multiplied alpha
When using pre-multiplied alpha, we see a color artifact in the final
output. Whether it is a missing something on our end or something in
overlaycomposition, we have not been able to pin point yet. Also, except
for overlaycomposition, nothing else in GStreamer is aware of or works
with pre-multiplied alpha at the moment.

Request non pre-multiplied output from canvex to fix this.