vcs/canvexoverlay: Allow canvex to be used as package from CanvexOverlay
Till now, we had to first invoke meson/ninja build in vcs/canvex and
then invoke cargo build in canvexoverlay to get things working.

With this change, using cc-rs we allow vcs/canvex side code to specified
as a package and be consumed as a library from rust. This allows us to
remove the intermediate step of generating the C API bindings and then
consuming that.

Now it is enough to run cargo build from canvexoverlay and cargo will
take care of building vcs/canvex as a library dependency.